Create tabelaverdade
The proposal is to develop a program that controls access to a system based on 
in three passwords, using propositional logic. The problem defines that access will be 
Awarded if both master passwords (P and Q) are correct or if the extra password (R) 
is correct, representing the logical condition (p∧q)∨r(p land q) lor r(p∧q)∨r.